requirements.txt
flask==3.0.2
opencv-python-headless
Pillow
pytesseract>=0.3
google-cloud-vision>=3.4
google-auth>=2.22
google-auth-oauthlib>=1.0
openai>=1.65.0